fix quadclipper in the case that the chop function fails
If the chopper fails, then we've hit some numerical edge-case, which indicates
that the quad is just barely crossing the edge, so to handle that, we just
clamp the Y values to the edge. This distorts the quad, but only in the case
when 99% of the quad will not be affected.