Better fix for changing the viewport in sample app.
Create a platform description and base the size on the SkWindow
in order to create a RenderTarget, which results in changing the
viewport.  This allows us to use only GR calls.

In the Android sample app, fix the problem where nothing is drawn
in the first frame, and default to drawing in GL.

Reviewed at http://codereview.appspot.com/4642043/