Fix a parenthesis bug.
SkGPipeCanvas::needOpBytes was being called with the wrong value due to a misplaced
parens in clipRect and clipPath. This can cause a crash if clip is called at just
the right (wrong) time. Instead of writing a boolean to the stream, I have added a
flag, which helps to avoid the parens problem.

Also renamed some flags from _DrawOpsFlag to _DrawOpFlag for consistency.

Lastly, added an assert that the size provided by the SkGPipeController is a multiple
of four.

Review URL: https://codereview.appspot.com/6453126