Revert r5520
Unfortunately, since we can only build ANGLE by including a gypi, and since
gyp cannot process includes conditionally, we *have* to have a build_angle.gypi
file in the third_party/externals/ANGLE/src directory.  We could do a hack
where we create an empty gypi if ANGLE is not checked out, but for now we just
check out ANGLE on all platforms.